Implement additional vtab functionality
Derives debug on the vtab IndexInfo struct as this aids in debugging vtable implementations.

Implements the set_idx_str, set_idx_flags, col_used, and collation methods on the vtab IndexInfo struct.

Add a close() method to the VTabCursor trait and call it from rust_close() so that the vtable
can clean up any state that it has related to the cursor, which in some cases may include unsafe
raw pointers that have been passed over the FFI boundary and need to be reclaimed to avoid leaking
memory.